+++ ephios/event_management/signup/__init__.py
@@ -1,0 +1,1 @@
+from .methods import *  # no-qa
